Validating Input
1. Modify PointOfInterestForCreationDto so that it:
a. requires the Name field and displays a custom error message
b. sets the max length for the name field to 50
b. sets the max length for the Description field to 200
PointOfInterestForCreationDto.cs: line 7-13

Note: There is no need to specify this check on the PointsOfInterestController body or any of its methods that use PointsOfInterestForCreationDto since the [ApiController] decorator will handle this.