Partially Updating a Resource
1. Add mapping to map PointOfInterest entity to PointOfInteresForUpdateDto model.
PointOfInterest.cs: line 12

2. Modify PartiallyUpdatePointOfInterest:
a. Check the repository asynchronously if the city exists
b. Load a single point of interest to pointOfInterestEntity and check if it is not null
c. Map pointOfInterestEntity to PointOfInterestForUpdateDto model and assign to pointOfInterestToPatch.
d. Apply the patchDocument to pointOfInterestToPatch.
e. Validate the model state.
f. Validate pointOfInterestToPatch.
g. Map the pointOfInterestToPatch with the updated point of interest back to the pointOfInterestEntity.
h. Save changes asynchronously.
i. Return a 204 no content response.
PointsOfInterestController.cs: line 126-162