Searching Through Resources
1. Modify the GetCitiesAsync overload on the ICityInfoRepository contract to take searchQuery as an input parameter when needed.
ICityInfoRepository.cs: line 8

2. Modify the implementation of the GetCitiesAsync overload so that:
a. it will check if a name and searchQuery is passed to it. If not, it executes the GetCitiesAsync without input parameters.
b. Assign cities as an iQueryable to the collection variable.
c. Check if a name is passed, and if so filters by name.
d. Check if a search query is passed, and if so retrieves cities which contains the search query.
CityInfoRepository.cs: line 21-46